pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "voxel"
description = "Base Library for building SPIM software"
license = { text = "MIT" }
requires-python = ">=3.11,<3.12"
authors = [
    { name = "Micah Woodard", email = "micah.woodard@alleninstitute.org" },
    { name = "Adam Glaser", email = "adam.glaser@alleninstitute.org" },
    { name = "Walter Mwaniki", email = "walter.mwaniki@alleninstitute.org" },
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    # base
    'ruamel-yaml >= 0.18.5',
    # hardware
    'nidaqmx >= 0.9.0',
    'pyserial >= 3.5',
    'pyusb >= 1.2.1',
    'PyVISA >= 1.14.1',
    'PyVISA-py >= 0.7.2',
    'pylablib >= 1.4.2',
    'tigerasi >= 0.0.26',
    # writers
    'tifffile >= 2024.1.30',
    'PyImarisWriter >= 0.7.0',
    'h5py >= 3.0.0', # for reading and writing HDF5 files
    'tqdm >= 4.66.5',
    # transfers
    'imohash >= 1.1.0',
    # essentials
    'numpy >= 1.23.0, < 1.24.0',
    'matplotlib >= 3.8.2',
    'scipy >= 1.12.0',
    'scikit-image >= 0.20.0',
    'scikit-tensor-py3',
    'tensorstore >= 0.1.56', # for reading and writing tensorstore files
    # utils
    'inflection >= 0.5.1',
    'gputools >= 0.2.14',
    'pyclesperanto >= 0.8.2', # for GPU acceleration
    'imohash >= 1.0.5',
    'psutil >= 5.9.8',
    'sympy >= 1.12.1', # used for symbolic math
]

[project.optional-dependencies]
dev = [
    'pytest >= 8.2.1',
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'furo'
]
#cameras = [
#    'egrabber >= 0.0.2',
#]
#lasers = [
#    'pycobolt @ git+https://github.com/cobolt-lasers/pycobolt.git',
#    'coherent_lasers @ git+https://github.com/AllenNeuralDynamics/coherent_lasers.git',
#    'vortran_laser @ git+https://github.com/AllenNeuralDynamics/vortran_laser.git',
#    'oxxius_laser @ git+https://github.com/AllenNeuralDynamics/oxxius_laser.git',
#    'obis_laser @ git+https://github.com/AllenNeuralDynamics/obis_laser.git',
#]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
version = { attr = "voxel.__version__" }

[tool.black]
line-length = 120
target_version = ['py36']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["voxel", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "from",
    "import",
    "pragma: no cover"
]
fail_under = 100

[tool.isort]
line_length = 120
profile = "black"

[tool.interrogate]
exclude = ["setup.py", "docs", "build"]
fail-under = 100
